% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gelnet.R
\name{adj2lapl}
\alias{adj2lapl}
\title{Generate a graph Laplacian}
\usage{
adj2lapl(A)
}
\arguments{
\item{A}{n-by-n adjacency matrix for a graph with n nodes}
}
\value{
The n-by-n Laplacian matrix of the graph
}
\description{
Generates a graph Laplacian from the graph adjacency matrix.
}
\details{
A graph Laplacian is defined as:
\eqn{ l_{i,j} = deg( v_i ) }, if \eqn{ i = j };
\eqn{ l_{i,j} = -1 }, if \eqn{ i \neq j } and \eqn{v_i} is adjacent to \eqn{v_j};
and \eqn{ l_{i,j} = 0 }, otherwise
}
\seealso{
\code{\link{adj2nlapl}}
}

