% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gelnet.R
\name{adj2nlapl}
\alias{adj2nlapl}
\title{Generate a normalized graph Laplacian}
\usage{
adj2nlapl(A)
}
\arguments{
\item{A}{n-by-n adjacency matrix for a graph with n nodes}
}
\value{
The n-by-n Laplacian matrix of the graph
}
\description{
Generates a normalized graph Laplacian from the graph adjacency matrix.
}
\details{
A normalized graph Laplacian is defined as:
\eqn{ l_{i,j} = 1 }, if \eqn{ i = j };
\eqn{ l_{i,j} = - 1 / \sqrt{ deg(v_i) deg(v_j) } }, if \eqn{ i \neq j } and \eqn{v_i} is adjacent to \eqn{v_j};
and \eqn{ l_{i,j} = 0 }, otherwise
}
\seealso{
\code{\link{adj2nlapl}}
}

