% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def.R
\name{gel_init}
\alias{gel_init}
\title{Initializer for GELnet models}
\usage{
gel_init(w_init = NULL, b_init = NULL)
}
\arguments{
\item{w_init}{p-by-1 vector of initial weight values}

\item{b_init}{scalar, initial value for the bias term}
}
\value{
An initializer that can be combined with a model definition using + operator
}
\description{
Defines initial values for model weights and the bias term
}
\details{
If an initializer is NULL, the values are computed automatically during training
}
