% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def.R
\name{gel_L1}
\alias{gel_L1}
\title{L1 regularizer}
\usage{
gel_L1(l1, d = NULL)
}
\arguments{
\item{l1}{coefficient for the L1-norm penalty}

\item{d}{p-by-1 vector of feature weights}
}
\value{
A regularizer definition that can be combined with a model definition using + operator
}
\description{
Defines an L1 regularizer with optional per-feature weights
}
\details{
The L1 regularization term is defined by
\deqn{ R1(w) = \lambda_1 \sum_j d_j |w_j| }
}
