% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def.R
\name{gelnet}
\alias{gelnet}
\title{GELnet model definition}
\usage{
gelnet(X)
}
\arguments{
\item{X}{n-by-p matrix of n samples in p dimensions}
}
\value{
A GELnet model definition
}
\description{
Starting building block for defining a GELnet model
}
