% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{gelnet_blr_opt}
\alias{gelnet_blr_opt}
\title{GELnet for logistic regression}
\usage{
gelnet_blr_opt(X, y, l1, l2, max_iter = 100L, eps = 1e-05, silent = FALSE,
  verbose = FALSE, balanced = FALSE, nonneg = FALSE, w_init = NULL,
  b_init = NULL, d = NULL, P = NULL, m = NULL)
}
\arguments{
\item{X}{n-by-p matrix of n samples in p dimensions}

\item{y}{n-by-1 vector of binary response labels (must be in {0,1})}

\item{l1}{coefficient for the L1-norm penalty}

\item{l2}{coefficient for the L2-norm penalty}

\item{eps}{convergence precision}

\item{silent}{set to TRUE to suppress run-time output to stdout (default: FALSE)}

\item{balanced}{boolean specifying whether the balanced model is being trained}

\item{nonneg}{set to TRUE to enforce non-negativity constraints on the weights (default: FALSE )}

\item{d}{p-by-1 vector of feature weights}

\item{P}{p-by-p feature association penalty matrix}

\item{m}{p-by-1 vector of translation coefficients}

\item{max.iter}{maximum number of iterations}

\item{w.init}{initial parameter estimate for the weights}

\item{b.init}{initial parameter estimate for the bias term}
}
\value{
A list with two elements:
\describe{
  \item{w}{p-by-1 vector of p model weights}
  \item{b}{scalar, bias term for the linear model}
}
}
\description{
Constructs a GELnet model for logistic regression using the Newton method.
}
\details{
The method operates by constructing iteratively re-weighted least squares approximations
of the log-likelihood loss function and then calling the linear regression routine
to solve those approximations. The least squares approximations are obtained via the Taylor series
expansion about the current parameter estimates.
}
\seealso{
\code{\link{gelnet.lin}}
}
