% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L1bin.R
\name{L1_ceiling}
\alias{L1_ceiling}
\title{The largest meaningful value of the L1 parameter}
\usage{
L1_ceiling(modeldef)
}
\arguments{
\item{modeldef}{model definition constructed through gelnet() arithmetic}
}
\value{
The largest meaningful value of the L1 parameter (i.e., the smallest value that yields a model with all zero weights)
}
\description{
Computes the smallest value of the LASSO coefficient L1 that leads to an
 all-zero weight vector for a given gelnet model
}
\details{
The cyclic coordinate descent updates the model weight \eqn{w_k} using a soft threshold operator
\eqn{ S( \cdot, \lambda_1 d_k ) } that clips the value of the weight to zero, whenever the absolute
value of the first argument falls below \eqn{\lambda_1 d_k}. From here, it is straightforward to compute
the smallest value of \eqn{\lambda_1}, such that all weights are clipped to zero.
}
