% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def.R
\name{model_blr}
\alias{model_blr}
\title{Binary logistic regression}
\usage{
model_blr(y, nonneg = FALSE, balanced = FALSE)
}
\arguments{
\item{y}{n-by-1 factor with two levels}

\item{nonneg}{set to TRUE to enforce non-negativity constraints on the weights (default: FALSE)}

\item{balanced}{boolean specifying whether the balanced model is being trained (default: FALSE)}
}
\value{
A GELnet task definition that can be combined with gelnet() output
}
\description{
Defines a binary logistic regression task
}
\details{
The binary logistic regression objective function is defined as
\deqn{ -\frac{1}{n} \sum_i y_i s_i - \log( 1 + \exp(s_i) ) + R(w) }
 where
\deqn{ s_i = w^T x_i + b }
}
