% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{perturb.gelnet}
\alias{perturb.gelnet}
\title{Perturbs a GELnet model}
\usage{
perturb.gelnet(model, i, delta)
}
\arguments{
\item{model}{The model to perturb}

\item{i}{Index of the coefficient to modify, or 0 for the bias term}

\item{delta}{The value to perturb by}
}
\value{
Modified GELnet model
}
\description{
Given a linear model, perturbs its i^th coefficient by delta
}
