% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/composite.R
\name{+.geldef}
\alias{+.geldef}
\title{Composition operator for GELnet model definition}
\usage{
\method{+}{geldef}(lhs, rhs)
}
\arguments{
\item{lhs}{left-hand side of composition (current chain)}

\item{rhs}{right-hand side of composition (new module)}
}
\description{
Composition operator for GELnet model definition
}
