% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def.R
\name{rglz_L2}
\alias{rglz_L2}
\title{L2 regularizer}
\usage{
rglz_L2(l2, P = NULL, m = NULL)
}
\arguments{
\item{l2}{coefficient for the L2-norm penalty}

\item{P}{p-by-p feature association penalty matrix}

\item{m}{p-by-1 vector of translation coefficients}
}
\value{
A regularizer definition that can be combined with a model definition using + operator
}
\description{
Defines an L2 regularizer with optional feature-feature penalties and translation coefficients
}
\details{
The L2 regularizer term is define by
\deqn{ R2(w) = \frac{\lambda_2}{2} (w-m)^T P (w-m) }
}
