% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def.R
\name{model_lin}
\alias{model_lin}
\title{Linear regression}
\usage{
model_lin(y, a = NULL, nonneg = FALSE, fix_bias = FALSE)
}
\arguments{
\item{y}{n-by-1 numeric vector of response values}

\item{a}{n-by-1 vector of sample weights}

\item{nonneg}{set to TRUE to enforce non-negativity constraints on the weights (default: FALSE)}

\item{fix_bias}{set to TRUE to prevent the bias term from being updated (default: FALSE)}
}
\value{
A GELnet task definition that can be combined with gelnet() output
}
\description{
Defines a linear regression task
}
\details{
The objective function is given by
\deqn{ \frac{1}{2n} \sum_i a_i (y_i - (w^T x_i + b))^2 + R(w) }
}
